+++ src/MC/JPetRawMCHit/JPetRawMCHit.cpp
@@ -34,11 +34,11 @@ void JPetRawMCHit::setMCDecayTreeIndex(int decayTreeIndex) { fMCDecayTreeIndex =
 
 void JPetRawMCHit::setMCVtxIndex(int vertexIndex) { fMCVtxIndex = vertexIndex; }
 
-void JPetRawMCHit::setPolarization(double polX, double polY, double polZ) { fPolarization.SetXYZ(polX, polY, polZ); }
+void JPetRawMCHit::setPolarization(float polX, float polY, float polZ) { fPolarization.SetXYZ(polX, polY, polZ); }
 
 void JPetRawMCHit::setPolarization(const TVector3& polarization) { fPolarization = polarization; }
 
-void JPetRawMCHit::setMomentum(double momX, double momY, double momZ) { fMomentum.SetXYZ(momX, momY, momZ); }
+void JPetRawMCHit::setMomentum(float momX, float momY, float momZ) { fMomentum.SetXYZ(momX, momY, momZ); }
 
 void JPetRawMCHit::setMomentum(const TVector3& momentum) { fMomentum = momentum; }
